@@ -140,7 +140,7 @@ require "coffee-script"
 require "builder"
 require "nokogiri"
 require "securerandom"
-require "aws-sdk"
+require "aws-sdk-s3"
 require "mimemagic"
 require "request_store"
 require "lokka/database"
