@@ -2,7 +2,7 @@ require File.expand_path(File.dirname(__FILE__) + '/spec_helper')
 
 describe '/admin/categories' do
   include_context 'admin login'
-  before { @category = Factory(:category) }
+  before { @category = create(:category) }
   after { Category.destroy }
 
   context 'GET /admin/categories' do
